third_party/WebKit/LayoutTests/fast/dom/MutationObserver/callback-arguments-expected.txt

Verifies that MutationObserver is passed to the callback as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS thisArgument is mutationObserver
PASS argument2 is mutationObserver
PASS successfullyParsed is true

TEST COMPLETE
